+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Fix list of targets to scrape or ignore.
+
 ## [4.20.5] - 2023-02-09
 
 ### Fixed
 
 - Manage etcd certificates differently between CAPI/Vintage. On Vintage, etcd certificates are binded via a volume. On CAPI, certificates are binded via a secret.
 - Pass the missing Provider property to `etcdcertificates.Config`
-
 
 ### Added